openwrt: init-script: fix mistake in pid-reader
shell vars must not be sourrounded bei single quotes,
otherwise they are not expanded. the error was hide by the
fact that there was already a fallback mechanism ("ask kernel
if pid is empty").

while we are at it, write global vars in BIG letters (style)
and only use ${VAR} syntax instead of $VAR if really needed